requirements.txt
# Usage: pip install -r requirements.txt

fastapi
uvicorn[standard]
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlmodel
google-cloud-translate
google-cloud-vision
google-cloud-storage
pillow